Separated service/server streams from server/browser streams.
You can now attach multiple browser windows to the same stream.  There is also some initial handling for disconnects/reconnects, but that still needs to be tested.

Currently there is a lot of console.log going on, and that will need to be removed eventually, but I want to commit this first working version so that it’s saved somewhere before I start trying to clean things up.